Fixup type

+++ schema/schema.go
@@ -399,7 +399,7 @@ type BranchChangesetSpec struct {
 	// Title description: The title of the changeset on the code host.
 	Title string `json:"title"`
 	// Version description: A field for versioning the payload.
-	Version float64 `json:"version,omitempty"`
+	Version int `json:"version,omitempty"`
 }
 type BrandAssets struct {
 	// Logo description: The URL to the image used on the homepage. This will replace the Sourcegraph logo on the homepage. Maximum width: 320px. We recommend using the following file formats: SVG, PNG
@@ -596,7 +596,7 @@ type ExistingChangesetSpec struct {
 	// ExternalID description: The ID that uniquely identifies the existing changeset on the code host
 	ExternalID string `json:"externalID"`
 	// Version description: A field for versioning the payload.
-	Version float64 `json:"version,omitempty"`
+	Version int `json:"version,omitempty"`
 }
 
 // ExpandedGitCommitDescription description: The Git commit to create with the changes.
@@ -772,7 +772,7 @@ type GitCommitDescription struct {
 	// Message description: The Git commit message.
 	Message string `json:"message"`
 	// Version description: A field for versioning the payload.
-	Version float64 `json:"version,omitempty"`
+	Version int `json:"version,omitempty"`
 }
 
 // GitHubApp description: The config options for Sourcegraph GitHub App.
